+++ crates/storage/src/disk/disk_manager/tests.rs
@@ -16,8 +16,8 @@ mod tests {
 
     #[test]
     fn read_write_page() {
-        let mut buf = [0u8; BUSTUB_PAGE_SIZE as usize];
-        let mut data = [0u8; BUSTUB_PAGE_SIZE as usize];
+        let mut buf = [0u8; BUSTUB_PAGE_SIZE];
+        let mut data = [0u8; BUSTUB_PAGE_SIZE];
         let tmp_dir = setup();
 
         let db_file = tmp_dir.path().join("test.db");
@@ -81,8 +81,8 @@ mod tests {
 
     #[test]
     fn read_write_page_unlimited_memory() {
-        let mut buf = [0u8; BUSTUB_PAGE_SIZE as usize];
-        let mut data = [0u8; BUSTUB_PAGE_SIZE as usize];
+        let mut buf = [0u8; BUSTUB_PAGE_SIZE];
+        let mut data = [0u8; BUSTUB_PAGE_SIZE];
         let tmp_dir = setup();
 
         let mut dm = DiskManagerUnlimitedMemory::new();
